requirements.txt
Django==3.2.8
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
asgiref==3.6.0
psycopg2==2.9.5
djoser==2.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.8.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
uritemplate==4.1.1
urllib3==1.26.14

